Moved QueryResult.kt to core
